config: s3: relax permissions to acl subpaths of '/system/s3-endpoint'
Currently the permissions to read/list s3 endpoint configurations
require Sys.Audit, editing requires Sys.Modify, both on the root path.

Switch the privileges to be less restrictive by allowing to set the
permissions on `/system/s3-endpoint/{id}` or `/system/s3-endpoint`.
By this, the permissions can be controlled with more fine tuning,
e.g. allowing to limit read access to a subset of the configured s3
endpoints.

Note, this is independent from the permissions to backup/modify
contents on a datastore. A user does not need to read the s3 endpoint
configuration to perform datastore operations.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250728075957.314427-4-c.ebner@proxmox.com