datastore: destroy: delete s3 in-use marker unconditionally
Currently, the in-use marker object for datastores backed by s3
backend is only removed together with the data when all data is
requested to be removed.

It the data should be preserved, the in-use marker should however
also be removed, allowing to reuse the datastore on the same or
a different host without force overwriting the marker.

Therefore, unconditionally clear the in-use marker from the S3
object store. In order to use the predefined marker filename
constant move it so it can be used during destruction as well.

Suggested-by: Lukas Wagner <l.wagner@proxmox.com>
Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250730152303.729778-1-c.ebner@proxmox.com