metric_collection: use ip link for determining the type of interfaces
Physical interfaces now can have arbitrary names with the introduction
of proxmox-network-interface pinning. In order to correctly report the
bandwidth metrics for PBS hosts, the method for determining whether an
interface is physical or not needs to be adjusted as well. Use the
adapted helper from proxmox-network-api to query the information about
interfaces and use that to determine the type of interface at every
site that uses the generated stats. To avoid spawning a new process
with every update loop, cache the initial query and use that
information.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Tested-by: Christian Ebner <c.ebner@proxmox.com>