fix #6665: never rename chunks on s3 client fetch errors
The chunk verification for the s3 backend always fetches the chunk
directly from the backend and verifies it based on the received
response.

Currently, when fetching the chunk contents failed in case of an
unrelated s3 client error, e.g. due to transient networking issues,
both the locally cached and the chunk in the object store, were
incorrectly marked as bad and renamed.

Instead, treat chunk fetching issues as error but do not treat the
chunk itself as bad.

Fixes: https://bugzilla.proxmox.com/show_bug.cgi?id=6665
Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250825131007.626777-8-c.ebner@proxmox.com