sync: pull: fix regression in resync of corrupt snapshots
In commit eea1176f ("pull: refactor source snapshot list filtering
logic") the filtering of snapshots to be synced when pulling from the
source datastore was split in preparation for correct pre-filtering
of the snapshot when the verified-only or encrypted-only flags have
been set.

This however introduced a regression, as previously also snapshots
older than the newest snapshot on the target were included in the
list, if the `resync_corrupt` parameter is set and the corresponding
target snapshot failed verification or the loading of the manifest to
check this failed, now however being filtered out unconditionally.

Fix this by keeping these snapshots in the list in case the
`resync_corrupt` flag is set.

Also rename the boolean flag from the incorrect `verified` to
`needs_resync`, in order to reflect the behaviour.

Fixes: eea1176f ("pull: refactor source snapshot list filtering logic")
Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Reviewed-by: Hannes Laimer <h.laimer@proxmox.com>
Tested-by: Hannes Laimer <h.laimer@proxmox.com>
Link: https://lore.proxmox.com/20250917102957.231963-1-c.ebner@proxmox.com