api: backup: avoid holding mutex and inline backup cleanup method
The method to cleanup backups due to failure (or because a benchmark)
currently acquires a mutex guard before removing the backup directory.

Since the S3 backend however needs to execute code in async context,
it is not deadlock save to hold the guard. However, the guard was
only acquired to set the state to finished, which is however not
necessary since the request handling future is already executed to
completion, so this has no further effect.

Therefore, drop the mutex locking altoghether and inline the now
trivial simple function call to the respective callsides.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/all/20250929080407.124362-4-c.ebner@proxmox.com