api: chunk upload: extend consistency checks for s3 backend
Currently, the chunk upload to the s3 backend does not perform
verification of the uploaded chunk prior to uploading it to the s3
backend, only before inserting in the cache. Further, digest
calculation and CRC computation were missing for these chunks.

Bring the consistency checks in-line with the ones performed on
regular datastores by reusing `UploadChunk`, making it reusable for
both cases by factoring out the chunk insert, which must only happen
after the backend upload on s3.

Reported-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>
Reviewed-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>
Signed-off-by: Christian Ebner <c.ebner@proxmox.com>