local store cache: replace evicted cache chunks instead of truncate
Evicted chunks have been truncated to size zero, keeping the chunk
file in place as in-use marker for the garbage collection but freeing
the chunk file contents. This can however lead to restores failing if
they already opened the chunk file for reading, as their contents are
now incomplete.

Fix this by instead replacing the chunk file with a zero sized file,
leaving the contents accessible for the already opened chunk readers.

By moving the logic from the local datastore cache to a helper method
on the chunk store, it is also assured that the operation is guarded
by the chunk store mutex lock to avoid races with chunk re-insert.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/all/20251008152125.849216-5-c.ebner@proxmox.com