local store cache: serve response fetched from s3 backend
If the chunk file was not found in the local datastore cache or the
chunk file was the zero marker, it is fetched from the s3 backend and
re-inserted.

Currently, the chunk is then re-read from the file, but that is not
required. Serve the still in-memory chunk data as result of the cacher
access call.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/all/20251008152125.849216-6-c.ebner@proxmox.com