local store cache: rework access cache fetching and insert logic
The local datastore cache has both, an in-memory LRU cache only
storing the digests and the chunk marks on the filesystem.
Chunks in the LRU cache have recently been accessed, therefore the
chunk contents are expected to be present in the local chunk file,
while no payload is present for evicted ones.

The current implementation relied on the cacher to fetch the chunk
data on cache misses, but required to re-read the chunk file after
the download, as the cacher interface does not allow to return a
payload value other than the one defined for the LRU cache, which is
however none.

Therefore, instead of using the LRU cache access method and in turn
the S3Cacher, rather try to access the local filesystem chunks
directly. They need to be accessed anyways, and further this avoids
possible races with download and insert, as now the held filehandle
either has a chunk with valid content and can bypass the backend,
or the chunk must be downloaded, serving the chunk from the fetched
data instead after inserting into the cache.

By unconditional re-insertion, it is assured that the chunk will be
marked as recently used in all cases and the least recently used one
is evicted.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/all/20251008152125.849216-8-c.ebner@proxmox.com