api/datastore: move group notes setting to the datastore
In an effort to abstract away the datastore backend related logic
from the api, move the set_group_notes to a helper method on the
datastore.

The new helper method now also assures that the exclusive lock on
the backup group is acquired before updating the notes, in order to
avoid possible race conditions, e.g. with backup group destruction.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/all/20251105122233.439382-3-c.ebner@proxmox.com