api: datastore: wait for active operations to clear before s3 refresh
Currently, the s3 refresh does not take into consideration already
ongoing active operations, only blocking new ones.

This will however lead to inconsistencies if there are ongoing read
or write operations. Therefore, actively wait for ongoing operatioins
to complete before running the actual refresh and keep the datastore
config locked so the maintenance mode cannot be altered.If an abort
was requested while waiting, clear the maintenance mode as well.

By adding this as helper, it can be reuse for datastore recreation as
well.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/all/20251113142214.719217-6-c.ebner@proxmox.com