sync: pull: add helper struct for download/reuse chunk tracking
In preparation for correctly tracking chunks downloaded during the
pull sync from the remote source and checking if the index file
currently being pulled is allowed to reuse a pre-existing local
chunk.

Therefore, instead of only tracking the chunk digest's of downloaded
chunks also store flags to signal whether the chunk is reusable by
this index file without download and if it has already been touched
during this group pull.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Reviewed-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>