sync: pull: fix possible chunk leak with malicious index file
Currently, pre-existing chunks present on the local sync targets
datastore can be reused by index files being downloaded from the pull
sync source, without validation that they actually are in possession
of the chunk data. In practice this is however unlikely, as this
would require to obtain the chunk digest by some other means, e.g.
brute forcing or guessing.

To avoid this, keep track of chunks which are reusable for the index
file because they have been present in an index file of a previous or
the same snapshot of the group, clearing this state for each group
being synced. Force re-download of others.

By also tracking if a chunk has already been touched, duplicate chunk
processing is avoided.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Reviewed-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>