client: fix flaky temporary value dropped while borrowed
I actually only saw that error in our rust-doc CI and could not really
reproduce it anywhere else, even though I tried to closely matched the
CI environment. Anyhow, its slightly shorter and easier to read, and
the compiler should be able to optimize this the same way as before,
not like it matters much in this cold-path here.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>