api: notification: correctly enumerate known push-sync-jobs
The second parameter of list_config_sync_jobs determines the direction
(either push, pull, or both) of jobs that should be returned. Using a
value of 'None' makes the implementation fall back to the default, which
is 'pull'. This led to only pull-style jobs being returned from the
'matcher-field-values' endpoint.

This issue likely existed since the introduction of push-style jobs.

Fixes: https://forum.proxmox.com/threads/170987
Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Reviewed-by: Christian Ebner <c.ebner@proxmox.com>
Tested-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20251215132733.216592-1-l.wagner@proxmox.com